Moved Input into a controller folder.
