+++ core/src/main/se/tevej/game/controller/input/TKeyBoard.java
@@ -1,0 +1,10 @@
+package main.se.tevej.game.controller.input;
+
+import main.se.tevej.game.controller.input.listenerInterfaces.OnTappedListener;
+
+
+public interface TKeyBoard {
+
+    void addTappedListener(OnTappedListener onTappedListener);
+
+}
