+++ core/src/main/se/tevej/game/controller/input/listenerInterfaces/OnMovedListener.java
@@ -1,0 +1,8 @@
+package main.se.tevej.game.controller.input.listenerInterfaces;
+
+import main.se.tevej.game.controller.input.TMouse;
+
+public interface OnMovedListener {
+    void onMoved(TMouse mouse);
+
+}
