+++ core/src/main/se/tevej/game/controller/input/listenerInterfaces/OnTappedListener.java
@@ -1,0 +1,8 @@
+package main.se.tevej.game.controller.input.listenerInterfaces;
+
+import main.se.tevej.game.controller.input.TKeyBoard;
+import main.se.tevej.game.controller.input.enums.TButton;
+
+public interface OnTappedListener {
+    void onTapped (TKeyBoard keyBoard, TButton button);
+}
