--- core/src/main/se/tevej/game/input/TKeyBoard.java
@@ -1,10 +1,0 @@
-package main.se.tevej.game.input;
-
-import main.se.tevej.game.input.listenerInterfaces.OnTappedListener;
-
-
-public interface TKeyBoard {
-
-    void addTappedListener(OnTappedListener onTappedListener);
-
-}
