--- core/src/main/se/tevej/game/input/listenerInterfaces/OnMovedListener.java
@@ -1,8 +1,0 @@
-package main.se.tevej.game.input.listenerInterfaces;
-
-import main.se.tevej.game.input.TMouse;
-
-public interface OnMovedListener {
-    void onMoved(TMouse mouse);
-
-}
