+++ core/src/main/se/tevej/game/libgdx/view/rendering/input/InputLibgdxAdapter.java
@@ -3,7 +3,7 @@ package main.se.tevej.game.libgdx.view.rendering.input;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.InputMultiplexer;
 import com.badlogic.gdx.InputProcessor;
-import main.se.tevej.game.input.enums.TButton;
+import main.se.tevej.game.controller.input.enums.TButton;
 
 import java.util.HashMap;
 import java.util.Map;
