+++ core/src/main/se/tevej/game/libgdx/view/rendering/input/InputLibgdxFactory.java
@@ -1,8 +1,8 @@
 package main.se.tevej.game.libgdx.view.rendering.input;
 
-import main.se.tevej.game.input.InputFactory;
-import main.se.tevej.game.input.TKeyBoard;
-import main.se.tevej.game.input.TMouse;
+import main.se.tevej.game.controller.input.InputFactory;
+import main.se.tevej.game.controller.input.TKeyBoard;
+import main.se.tevej.game.controller.input.TMouse;
 
 public class InputLibgdxFactory implements InputFactory {
